+++ hydra-node/test/Hydra/Chain/Direct/Contract/Commit.hs
@@ -10,9 +10,11 @@ import Hydra.Chain.Direct.TxSpec ()
 
 import qualified Cardano.Api.UTxO as UTxO
 import Data.Maybe (fromJust)
+import Hydra.Chain.Direct.Contract.Gen (genMintedOrBurnedValue)
 import Hydra.Chain.Direct.Contract.Mutation (
   Mutation (..),
   SomeMutation (..),
+  changeMintedTokens,
   replacePolicyIdWith,
  )
 import qualified Hydra.Chain.Direct.Fixture as Fixture
@@ -80,6 +82,8 @@ data CommitMutation
   | MutateRequiredSigner
   | -- | Change the policy Id of the ST and PTs both in input and output
     MutateHeadId
+  | -- | Minting or burning of the tokens should not be possible in v_initial when checking the commit
+    MutateTokenMintingOrBurning
   deriving (Generic, Show, Enum, Bounded)
 
 genCommitMutation :: (Tx, UTxO) -> Gen SomeMutation
@@ -109,6 +113,8 @@ genCommitMutation (tx, _utxo) =
                 (toUTxOContext $ replacePolicyIdWith Fixture.testPolicyId otherHeadId healthyInitialTxOut)
                 (Just $ toScriptData $ Initial.ViaCommit $ Just $ toPlutusTxOutRef committedTxIn)
             ]
+    , SomeMutation (Just "minting or burning is forbidden") MutateTokenMintingOrBurning
+        <$> (changeMintedTokens tx =<< genMintedOrBurnedValue)
     ]
  where
   TxOut{txOutValue = commitOutputValue} = commitTxOut